Generalize post update query handler + model
