+++ API/src/Factories/HandlerFactory.php
@@ -681,14 +681,6 @@ namespace Timetabio\API\Factories
             );
         }
 
-        public function createArchivePostQueryHandler(): \Timetabio\API\Handlers\Post\Post\Archive\QueryHandler
-        {
-            return new \Timetabio\API\Handlers\Post\Post\Archive\QueryHandler(
-                $this->getMasterFactory()->createFetchPostInfoQuery(),
-                $this->getMasterFactory()->createFeedAccessControl()
-            );
-        }
-
         public function createArchivePostCommandHandler(): \Timetabio\API\Handlers\Post\Post\Archive\CommandHandler
         {
             return new \Timetabio\API\Handlers\Post\Post\Archive\CommandHandler(
@@ -700,6 +692,14 @@ namespace Timetabio\API\Factories
         {
             return new \Timetabio\API\Handlers\Post\Post\Restore\CommandHandler(
                 $this->getMasterFactory()->createRestorePostCommand()
+            );
+        }
+
+        public function createUpdatePostQueryHandler(): \Timetabio\API\Handlers\UpdatePostQueryHandler
+        {
+            return new \Timetabio\API\Handlers\UpdatePostQueryHandler(
+                $this->getMasterFactory()->createFetchPostInfoQuery(),
+                $this->getMasterFactory()->createFeedAccessControl()
             );
         }
     }
