+++ API/src/Handlers/UpdatePostQueryHandler.php
@@ -1,23 +1,23 @@
 <?php
 /**
- * Copyright (c) 2016 Ruben Schmidmeister <ruben.schmidmeister@icloud.com>
+ * Copyright (c) 2017 Ruben Schmidmeister <ruben.schmidmeister@icloud.com>
  *
  * This program is free software. It comes without any warranty, to
  * the extent permitted by applicable law. You can redistribute it
  * and/or modify it under the terms of the GNU Affero General Public License,
  * version 3, as published by the Free Software Foundation.
  */
-namespace Timetabio\API\Handlers\Post\Post\Archive
+namespace Timetabio\API\Handlers
 {
     use Timetabio\API\Access\AccessControl\FeedAccessControl;
     use Timetabio\API\Exceptions\Forbidden;
     use Timetabio\API\Exceptions\NotFound;
-    use Timetabio\API\Models\Post\ArchiveModel;
+    use Timetabio\API\Models\Post\PostModel;
     use Timetabio\API\Queries\Post\FetchPostInfoQuery;
     use Timetabio\Framework\Handlers\QueryHandlerInterface;
     use Timetabio\Framework\Models\AbstractModel;
 
-    class QueryHandler implements QueryHandlerInterface
+    class UpdatePostQueryHandler implements QueryHandlerInterface
     {
         /**
          * @var FetchPostInfoQuery
@@ -37,7 +37,7 @@ namespace Timetabio\API\Handlers\Post\Post\Archive
 
         public function execute(AbstractModel $model)
         {
-            /** @var ArchiveModel $model */
+            /** @var PostModel $model */
 
             $accessToken = $model->getAccessToken();
             $post = $this->fetchPostInfoQuery->execute($model->getPostId());
