--- API/src/Models/Post/ArchiveModel.php
@@ -1,29 +1,0 @@
-<?php
-/**
- * Copyright (c) 2016 Ruben Schmidmeister <ruben.schmidmeister@icloud.com>
- *
- * This program is free software. It comes without any warranty, to
- * the extent permitted by applicable law. You can redistribute it
- * and/or modify it under the terms of the GNU Affero General Public License,
- * version 3, as published by the Free Software Foundation.
- */
-namespace Timetabio\API\Models\Post
-{
-    class ArchiveModel extends PostModel
-    {
-        /**
-         * @var array
-         */
-        private $postInfo;
-
-        public function getPostInfo(): array
-        {
-            return $this->postInfo;
-        }
-
-        public function setPostInfo(array $postInfo)
-        {
-            $this->postInfo = $postInfo;
-        }
-    }
-}
